Expose touchpad settings if synaptics is in use
Upstream is aggressively pushing for libinput over synaptics. However that
creates a factually incorrect situation where gnome-control-center thinks
there is no touchpad unless you're using the xserver-xorg-input-libinput.
This patch adds support for both drivers, removing the code that would
hide the touchpad settings when xserver-xorg-input-synaptics is in use.

This patch is only useful in the GNOME on Xorg session.
In GNOME on Wayland, libinput is used anyway.

This patch requires a companion patch in mutter which GNOME has rejected.

https://gitlab.gnome.org/GNOME/mutter/merge_requests/37