network: always set WirelessEnabled NM props on wifi toggle
this is an interesting check. it essentially says only set WirelessEnabled to a state if the state is true. but if false state is never allowed but then we never need to touch WirelessEnabled since the value will never change and wifi can never be disabled either.
I don't understand what the point of this check is, but the logic here is wrong. in GNOME 46 and 47 this was correct but now it isn't even on stock GCC.

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>