a11y: fix accessibility when choosing a background
The background selector in the Appearance panel shows a set of
stock backgrounds. But when using the keyboard to select one and
a screen reader, nothing is described when any of the pictures
is selected. This is a regression from older versions.

This bug has been fixed upstream, with

https://gitlab.gnome.org/GNOME/gnome-control-center/-/merge_requests/2828

This patch backports this patch to Ubuntu 24.04

Fix #2114995